requirements.txt
paramiko==3.5.0
PyYAML==6.0.2
typer==0.12.5
rich==13.9.4
pytest==8.3.3
pytest-cov==5.0.0

